[SYCL] Break include's cycle
