+++ sycl/include/CL/sycl/detail/cg_types.hpp
@@ -1,0 +1,307 @@
+#pragma once
+
+#include <CL/sycl/detail/kernel_desc.hpp>
+#include <CL/sycl/detail/host_profiling_info.hpp>
+#include <CL/sycl/group.hpp>
+#include <CL/sycl/id.hpp>
+#include <CL/sycl/interop_handler.hpp>
+#include <CL/sycl/kernel.hpp>
+#include <CL/sycl/nd_item.hpp>
+#include <CL/sycl/range.hpp>
+
+__SYCL_INLINE_NAMESPACE(cl) {
+namespace sycl {
+namespace detail {
+
+// The structure represents kernel argument.
+class ArgDesc {
+public:
+  ArgDesc(cl::sycl::detail::kernel_param_kind_t Type, void *Ptr, int Size,
+          int Index)
+      : MType(Type), MPtr(Ptr), MSize(Size), MIndex(Index) {}
+
+  cl::sycl::detail::kernel_param_kind_t MType;
+  void *MPtr;
+  int MSize;
+  int MIndex;
+};
+
+// The structure represents NDRange - global, local sizes, global offset and
+// number of dimensions.
+class NDRDescT {
+  // The method initializes all sizes for dimensions greater than the passed one
+  // to the default values, so they will not affect execution.
+  void setNDRangeLeftover(int Dims_) {
+    for (int I = Dims_; I < 3; ++I) {
+      GlobalSize[I] = 1;
+      LocalSize[I] = LocalSize[0] ? 1 : 0;
+      GlobalOffset[I] = 0;
+      NumWorkGroups[I] = 0;
+    }
+  }
+
+public:
+  NDRDescT()
+      : GlobalSize{0, 0, 0}, LocalSize{0, 0, 0}, NumWorkGroups{0, 0, 0} {}
+
+  template <int Dims_> void set(sycl::range<Dims_> NumWorkItems) {
+    for (int I = 0; I < Dims_; ++I) {
+      GlobalSize[I] = NumWorkItems[I];
+      LocalSize[I] = 0;
+      GlobalOffset[I] = 0;
+      NumWorkGroups[I] = 0;
+    }
+    setNDRangeLeftover(Dims_);
+    Dims = Dims_;
+  }
+
+  // Initializes this ND range descriptor with given range of work items and
+  // offset.
+  template <int Dims_>
+  void set(sycl::range<Dims_> NumWorkItems, sycl::id<Dims_> Offset) {
+    for (int I = 0; I < Dims_; ++I) {
+      GlobalSize[I] = NumWorkItems[I];
+      LocalSize[I] = 0;
+      GlobalOffset[I] = Offset[I];
+      NumWorkGroups[I] = 0;
+    }
+    setNDRangeLeftover(Dims_);
+    Dims = Dims_;
+  }
+
+  template <int Dims_> void set(sycl::nd_range<Dims_> ExecutionRange) {
+    for (int I = 0; I < Dims_; ++I) {
+      GlobalSize[I] = ExecutionRange.get_global_range()[I];
+      LocalSize[I] = ExecutionRange.get_local_range()[I];
+      GlobalOffset[I] = ExecutionRange.get_offset()[I];
+      NumWorkGroups[I] = 0;
+    }
+    setNDRangeLeftover(Dims_);
+    Dims = Dims_;
+  }
+
+  void set(int Dims_, sycl::nd_range<3> ExecutionRange) {
+    for (int I = 0; I < Dims_; ++I) {
+      GlobalSize[I] = ExecutionRange.get_global_range()[I];
+      LocalSize[I] = ExecutionRange.get_local_range()[I];
+      GlobalOffset[I] = ExecutionRange.get_offset()[I];
+      NumWorkGroups[I] = 0;
+    }
+    setNDRangeLeftover(Dims_);
+    Dims = Dims_;
+  }
+
+  template <int Dims_> void setNumWorkGroups(sycl::range<Dims_> N) {
+    for (int I = 0; I < Dims_; ++I) {
+      GlobalSize[I] = 0;
+      // '0' is a mark to adjust before kernel launch when there is enough info:
+      LocalSize[I] = 0;
+      GlobalOffset[I] = 0;
+      NumWorkGroups[I] = N[I];
+    }
+    setNDRangeLeftover(Dims_);
+    Dims = Dims_;
+  }
+
+  sycl::range<3> GlobalSize;
+  sycl::range<3> LocalSize;
+  sycl::id<3> GlobalOffset;
+  /// Number of workgroups, used to record the number of workgroups from the
+  /// simplest form of parallel_for_work_group. If set, all other fields must be
+  /// zero
+  sycl::range<3> NumWorkGroups;
+  size_t Dims;
+};
+
+// The pure virtual class aimed to store lambda/functors of any type.
+class HostKernelBase {
+public:
+  // The method executes lambda stored using NDRange passed.
+  virtual void call(const NDRDescT &NDRDesc, HostProfilingInfo *HPI) = 0;
+  // Return pointer to the lambda object.
+  // Used to extract captured variables.
+  virtual char *getPtr() = 0;
+  virtual ~HostKernelBase() = default;
+};
+
+class InteropTask {
+  std::function<void(cl::sycl::interop_handler)> MFunc;
+
+public:
+  InteropTask(function_class<void(cl::sycl::interop_handler)> Func)
+      : MFunc(Func) {}
+  void call(cl::sycl::interop_handler &h) { MFunc(h); }
+};
+
+class HostTask {
+  std::function<void()> MHostTask;
+
+public:
+  HostTask() : MHostTask([]() {}) {}
+  HostTask(std::function<void()> &&Func) : MHostTask(Func) {}
+
+  void call() { MHostTask(); }
+};
+
+// Class which stores specific lambda object.
+template <class KernelType, class KernelArgType, int Dims>
+class HostKernel : public HostKernelBase {
+  using IDBuilder = sycl::detail::Builder;
+  KernelType MKernel;
+
+public:
+  HostKernel(KernelType Kernel) : MKernel(Kernel) {}
+  void call(const NDRDescT &NDRDesc, HostProfilingInfo *HPI) override {
+    // adjust ND range for serial host:
+    NDRDescT AdjustedRange = NDRDesc;
+
+    if (NDRDesc.GlobalSize[0] == 0 && NDRDesc.NumWorkGroups[0] != 0) {
+      // This is a special case - NDRange information is not complete, only the
+      // desired number of work groups is set by the user. Choose work group
+      // size (LocalSize), calculate the missing NDRange characteristics
+      // needed to invoke the kernel and adjust the NDRange descriptor
+      // accordingly. For some devices the work group size selection requires
+      // access to the device's properties, hence such late "adjustment".
+      range<3> WGsize{1, 1, 1}; // no better alternative for serial host?
+      AdjustedRange.set(NDRDesc.Dims,
+                        nd_range<3>(NDRDesc.NumWorkGroups * WGsize, WGsize));
+    }
+    // If local size for host is not set explicitly, let's adjust it to 1,
+    // so nd_range_error for zero local size is not thrown.
+    if (AdjustedRange.LocalSize[0] == 0)
+      for (size_t I = 0; I < AdjustedRange.Dims; ++I)
+        AdjustedRange.LocalSize[I] = 1;
+    if (HPI)
+      HPI->start();
+    runOnHost(AdjustedRange);
+    if (HPI)
+      HPI->end();
+  }
+
+  char *getPtr() override { return reinterpret_cast<char *>(&MKernel); }
+
+  template <class ArgT = KernelArgType>
+  typename std::enable_if<std::is_same<ArgT, void>::value>::type
+  runOnHost(const NDRDescT &) {
+    MKernel();
+  }
+
+  template <class ArgT = KernelArgType>
+  typename std::enable_if<std::is_same<ArgT, sycl::id<Dims>>::value>::type
+  runOnHost(const NDRDescT &NDRDesc) {
+    sycl::range<Dims> Range(InitializedVal<Dims, range>::template get<0>());
+    for (int I = 0; I < Dims; ++I)
+      Range[I] = NDRDesc.GlobalSize[I];
+
+    detail::NDLoop<Dims>::iterate(
+        Range, [&](const sycl::id<Dims> &ID) { MKernel(ID); });
+  }
+
+  template <class ArgT = KernelArgType>
+  typename std::enable_if<
+      std::is_same<ArgT, item<Dims, /*Offset=*/false>>::value>::type
+  runOnHost(const NDRDescT &NDRDesc) {
+    sycl::id<Dims> ID;
+    sycl::range<Dims> Range(InitializedVal<Dims, range>::template get<0>());
+    for (int I = 0; I < Dims; ++I)
+      Range[I] = NDRDesc.GlobalSize[I];
+
+    detail::NDLoop<Dims>::iterate(Range, [&](const sycl::id<Dims> ID) {
+      sycl::item<Dims, /*Offset=*/false> Item =
+          IDBuilder::createItem<Dims, false>(Range, ID);
+      MKernel(Item);
+    });
+  }
+
+  template <class ArgT = KernelArgType>
+  typename std::enable_if<
+      std::is_same<ArgT, item<Dims, /*Offset=*/true>>::value>::type
+  runOnHost(const NDRDescT &NDRDesc) {
+    sycl::range<Dims> Range(InitializedVal<Dims, range>::template get<0>());
+    sycl::id<Dims> Offset;
+    for (int I = 0; I < Dims; ++I) {
+      Range[I] = NDRDesc.GlobalSize[I];
+      Offset[I] = NDRDesc.GlobalOffset[I];
+    }
+
+    detail::NDLoop<Dims>::iterate(Range, [&](const sycl::id<Dims> &ID) {
+      sycl::id<Dims> OffsetID = ID + Offset;
+      sycl::item<Dims, /*Offset=*/true> Item =
+          IDBuilder::createItem<Dims, true>(Range, OffsetID, Offset);
+      MKernel(Item);
+    });
+  }
+
+  template <class ArgT = KernelArgType>
+  typename std::enable_if<std::is_same<ArgT, nd_item<Dims>>::value>::type
+  runOnHost(const NDRDescT &NDRDesc) {
+    sycl::range<Dims> GroupSize(InitializedVal<Dims, range>::template get<0>());
+    for (int I = 0; I < Dims; ++I) {
+      if (NDRDesc.LocalSize[I] == 0 ||
+          NDRDesc.GlobalSize[I] % NDRDesc.LocalSize[I] != 0)
+        throw sycl::nd_range_error("Invalid local size for global size",
+                                   PI_INVALID_WORK_GROUP_SIZE);
+      GroupSize[I] = NDRDesc.GlobalSize[I] / NDRDesc.LocalSize[I];
+    }
+
+    sycl::range<Dims> LocalSize(InitializedVal<Dims, range>::template get<0>());
+    sycl::range<Dims> GlobalSize(
+        InitializedVal<Dims, range>::template get<0>());
+    sycl::id<Dims> GlobalOffset;
+    for (int I = 0; I < Dims; ++I) {
+      GlobalOffset[I] = NDRDesc.GlobalOffset[I];
+      LocalSize[I] = NDRDesc.LocalSize[I];
+      GlobalSize[I] = NDRDesc.GlobalSize[I];
+    }
+
+    detail::NDLoop<Dims>::iterate(GroupSize, [&](const id<Dims> &GroupID) {
+      sycl::group<Dims> Group = IDBuilder::createGroup<Dims>(
+          GlobalSize, LocalSize, GroupSize, GroupID);
+
+      detail::NDLoop<Dims>::iterate(LocalSize, [&](const id<Dims> &LocalID) {
+        id<Dims> GlobalID = GroupID * LocalSize + LocalID + GlobalOffset;
+        const sycl::item<Dims, /*Offset=*/true> GlobalItem =
+            IDBuilder::createItem<Dims, true>(GlobalSize, GlobalID,
+                                              GlobalOffset);
+        const sycl::item<Dims, /*Offset=*/false> LocalItem =
+            IDBuilder::createItem<Dims, false>(LocalSize, LocalID);
+        const sycl::nd_item<Dims> NDItem =
+            IDBuilder::createNDItem<Dims>(GlobalItem, LocalItem, Group);
+        MKernel(NDItem);
+      });
+    });
+  }
+
+  template <typename ArgT = KernelArgType>
+  enable_if_t<std::is_same<ArgT, cl::sycl::group<Dims>>::value>
+  runOnHost(const NDRDescT &NDRDesc) {
+    sycl::range<Dims> NGroups(InitializedVal<Dims, range>::template get<0>());
+
+    for (int I = 0; I < Dims; ++I) {
+      if (NDRDesc.LocalSize[I] == 0 ||
+          NDRDesc.GlobalSize[I] % NDRDesc.LocalSize[I] != 0)
+        throw sycl::nd_range_error("Invalid local size for global size",
+                                   PI_INVALID_WORK_GROUP_SIZE);
+      NGroups[I] = NDRDesc.GlobalSize[I] / NDRDesc.LocalSize[I];
+    }
+
+    sycl::range<Dims> LocalSize(InitializedVal<Dims, range>::template get<0>());
+    sycl::range<Dims> GlobalSize(
+        InitializedVal<Dims, range>::template get<0>());
+    for (int I = 0; I < Dims; ++I) {
+      LocalSize[I] = NDRDesc.LocalSize[I];
+      GlobalSize[I] = NDRDesc.GlobalSize[I];
+    }
+    detail::NDLoop<Dims>::iterate(NGroups, [&](const id<Dims> &GroupID) {
+      sycl::group<Dims> Group =
+          IDBuilder::createGroup<Dims>(GlobalSize, LocalSize, NGroups, GroupID);
+      MKernel(Group);
+    });
+  }
+
+  ~HostKernel() = default;
+};
+
+} // namespace detail
+} // namespace sycl
+} // __SYCL_INLINE_NAMESPACE(cl)
