+++ sycl/include/CL/sycl/interop_handler.hpp
@@ -1,0 +1,64 @@
+#pragma once
+
+#include <CL/sycl/accessor.hpp>
+
+__SYCL_INLINE_NAMESPACE(cl) {
+namespace sycl {
+
+// Interoperability handler
+//
+class interop_handler {
+  // Make accessor class friend to access the detail mem objects
+  template <typename DataT, int Dims, access::mode AccMode,
+            access::target AccTarget, access::placeholder isPlaceholder>
+  friend class accessor;
+
+public:
+  using QueueImplPtr = std::shared_ptr<detail::queue_impl>;
+  using ReqToMem = std::pair<detail::Requirement *, pi_mem>;
+
+  interop_handler(std::vector<ReqToMem> MemObjs, QueueImplPtr Queue)
+      : MQueue(std::move(Queue)), MMemObjs(std::move(MemObjs)) {}
+
+  template <backend BackendName = backend::opencl>
+  auto get_queue() const -> typename interop<BackendName, queue>::type {
+    return reinterpret_cast<typename interop<BackendName, queue>::type>(
+        GetNativeQueue());
+  }
+
+  template <backend BackendName = backend::opencl, typename DataT, int Dims,
+            access::mode AccessMode, access::target AccessTarget,
+            access::placeholder IsPlaceholder = access::placeholder::false_t>
+  auto get_mem(accessor<DataT, Dims, AccessMode, AccessTarget,
+                        access::placeholder::false_t>
+                   Acc) const ->
+      typename interop<BackendName,
+                       accessor<DataT, Dims, AccessMode, AccessTarget,
+                                access::placeholder::false_t>>::type {
+    detail::AccessorBaseHost *AccBase = (detail::AccessorBaseHost *)&Acc;
+    return getMemImpl<BackendName, DataT, Dims, AccessMode, AccessTarget,
+                      access::placeholder::false_t>(
+        detail::getSyclObjImpl(*AccBase).get());
+  }
+
+private:
+  QueueImplPtr MQueue;
+  std::vector<ReqToMem> MMemObjs;
+
+  template <backend BackendName, typename DataT, int Dims,
+            access::mode AccessMode, access::target AccessTarget,
+            access::placeholder IsPlaceholder>
+  auto getMemImpl(detail::Requirement *Req) const -> typename interop<
+      BackendName,
+      accessor<DataT, Dims, AccessMode, AccessTarget, IsPlaceholder>>::type {
+    return (typename interop<BackendName,
+                             accessor<DataT, Dims, AccessMode, AccessTarget,
+                                      IsPlaceholder>>::type)GetNativeMem(Req);
+  }
+
+  __SYCL_EXPORT pi_native_handle GetNativeMem(detail::Requirement *Req) const;
+  __SYCL_EXPORT pi_native_handle GetNativeQueue() const;
+};
+
+} // namespace sycl
+} // __SYCL_INLINE_NAMESPACE(cl)
